Revert "fix lint"
This reverts commit 269acd1715a93d59f61286d336afa9a03f07f805.

+++ cocos/2d/assembler/sprite/sliced.ts
@@ -49,7 +49,7 @@ export const sliced: IAssembler = {
         renderData.resize(16, 54);
         this.QUAD_INDICES = new Uint16Array(54);
         this.createQuadIndices(4, 4);
-        renderData.chunk.setIndexBuffer(this.QUAD_INDICES as Uint16Array);
+        renderData.chunk.setIndexBuffer(this.QUAD_INDICES);
         return renderData;
     },
